PP-5170: capture charge on sqs captrue message
Use `CaptureService` `doCapture` to capture charge based on external ID
from SQS message.

+++ src/main/java/uk/gov/pay/connector/queue/CaptureQueue.java
@@ -5,6 +5,7 @@ import com.google.gson.GsonBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.connector.app.ConnectorConfiguration;
+import uk.gov.pay.connector.gateway.CaptureResponse;
 import uk.gov.pay.connector.paymentprocessor.service.CardCaptureMessageProcess;
 import uk.gov.pay.connector.queue.sqs.SqsQueueService;
 
@@ -48,8 +49,17 @@ public class CaptureQueue {
         List<QueueMessage> captureMessages = sqsQueueService.receiveMessages(this.captureQueueUrl, CAPTURE_MESSAGE_ATTRIBUTE_NAME);
         for (QueueMessage message: captureMessages) {
             try {
-                cardCaptureMessageProcess.runCapture(message);
-                sqsQueueService.deleteMessage(this.captureQueueUrl, message);
+                CaptureResponse gatewayResponse = cardCaptureMessageProcess.runCapture(message);
+                
+                if (gatewayResponse.isSuccessful()) {
+                    sqsQueueService.deleteMessage(this.captureQueueUrl, message);
+                } else {
+                    logger.info(
+                            "Failed to capture [messageBody={}] due to: {}",
+                            message.getMessageBody(),
+                            gatewayResponse.getError().get().getMessage()
+                    );
+                }
             } catch (Exception e) {
                 logger.warn("Error capturing charge from SQS message [{}]", e);
             }
